+++ CHANGELOG.md
@@ -15,18 +15,30 @@
 ## Compatibility Notes
 
 - [eslint-config] Remove `brace-style` rule for compatibility with Prettier 2.3.
+- [openapi-generator] Change the basis for directory, package, and service names, when generating clients. If not specified otherwise, the default is based on the directory name instead of the service name.
+- [core] Deprecate JWT related interfaces in favor of the interfaces provided by the `jsonwebtoken` library.
+The following interfaces were deprecated:
+  - `JWTHeader` (use `JwtHeader` instead, the property `typ` is now optional)
+  - `JWTPayload` (use `JwtPayload` instead)
+  - `CompleteDecodedJWT` (use `Jwt` instead)
+  - `RegisteredJWTClaims`
+  - `RegisteredJWTClaimsBasic`
+  - `RegisteredJWTClaimsTenant`
+  - `RegisteredJWTClaimsUser`
 
 ## New Functionality
 
--
+- [openapi-generator] Add the `-c/--config` command line option to specify options through a configuration file instead of on the command line.
 
 ## Improvements
 
--
+- [openapi-generator] Improve the error message for invalid or unsupported path patterns.
+- [core] Improve types in convenience functions for JWT access. See "Compatibility Notes" for details.
 
 ## Fixed Issues
 
--
+- [core] Fix type error to allow filtering on one-to-many navigation properties in lambda expressions.
+- [openapi-generator] Base uniqueness check for directory names on directory names in `optionsPerService` instead of the human readable service name.
 
 
 # 1.45.0